+++ frontend/app/src/components/table-filter/TableFilter.vue
@@ -170,7 +170,7 @@ const filteredMatchers = computed<SearchMatcher<any>[]>(() => {
 
   const searchToken = getTextToken(searchVal);
   return filteredByUsedKeys
-    .filter(({ key }) => getTextToken(key).includes(searchVal))
+    .filter(({ key }) => getTextToken(key).includes(searchToken))
     .sort((a, b) => compareTextByKeyword(getTextToken(a.key), getTextToken(b.key), searchToken));
 });